+++ server/src/main/kotlin/org/kryptonmc/krypton/command/CommandScope.kt
@@ -2,9 +2,13 @@ package org.kryptonmc.krypton.command
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.asCoroutineDispatcher
+import org.kryptonmc.krypton.concurrent.NamedThreadFactory
 import java.util.concurrent.Executors
 
 object CommandScope : CoroutineScope {
 
-    override val coroutineContext = Executors.newFixedThreadPool(8).asCoroutineDispatcher()
+    override val coroutineContext = Executors.newFixedThreadPool(
+        8,
+        NamedThreadFactory("Command Handler #%d")
+    ).asCoroutineDispatcher()
 }