+++ server/src/main/kotlin/org/kryptonmc/krypton/entity/entities/KryptonPlayer.kt
@@ -1,5 +1,7 @@
 package org.kryptonmc.krypton.entity.entities
 
+import net.kyori.adventure.audience.MessageType
+import net.kyori.adventure.identity.Identity
 import net.kyori.adventure.text.Component
 import org.kryptonmc.krypton.api.entity.Abilities
 import org.kryptonmc.krypton.api.entity.entities.Player
@@ -8,11 +10,13 @@ import org.kryptonmc.krypton.api.world.Gamemode
 import org.kryptonmc.krypton.api.world.Location
 import org.kryptonmc.krypton.api.world.scoreboard.Scoreboard
 import org.kryptonmc.krypton.command.KryptonSender
+import org.kryptonmc.krypton.packet.out.play.chat.PacketOutChat
+import org.kryptonmc.krypton.session.Session
 import java.net.InetSocketAddress
 import java.util.*
 
 class KryptonPlayer(
-    val id: Int,
+    val session: Session,
     override val address: InetSocketAddress = InetSocketAddress("127.0.0.1", 1)
 ) : Player, KryptonSender() {
 
@@ -39,4 +43,8 @@ class KryptonPlayer(
     override lateinit var locale: Locale
 
     var gamemode = Gamemode.SURVIVAL
+
+    override fun sendMessage(source: Identity, message: Component, type: MessageType) {
+        session.sendPacket(PacketOutChat(message, type, source.uuid()))
+    }
 }